+++ src/core/index.ts
@@ -1,4 +1,4 @@
 export * from "./core.js";
 export { applyElement } from "./jsx/common.js";
-export * from "./overwrite-jsx.js";
+export * from "./override-jsx.js";
 export * from "./view-jsx.js";