+++ src/core/internals/set-style.ts
@@ -6,16 +6,16 @@ type StyleHandler = (name: string, value: unknown) => void;
 function watchStyle(value: StyleValue, handler: StyleHandler): void {
 	watch(value, value => {
 		if (Array.isArray(value)) {
-			const overwrites: string[][] = [];
+			const overrides: string[][] = [];
 			for (let i = value.length - 1; i >= 0; i--) {
 				const self: string[] = [];
-				overwrites[i] = self;
+				overrides[i] = self;
 				watchStyle(value[i], (name, value) => {
 					if (!self.includes(name)) {
 						self.push(name);
 					}
-					for (let o = i + 1; o < overwrites.length; o++) {
-						if (overwrites[o].includes(name)) {
+					for (let o = i + 1; o < overrides.length; o++) {
+						if (overrides[o].includes(name)) {
 							return;
 						}
 					}